+++ src/data/id/mod.rs
@@ -3,13 +3,14 @@
 //! Data identifiers.
 //
 
+mod marker; // zero-cost generic ids
 mod pin; // pinned memory-based ids
 mod res; // type-safe resource ids
 mod seq; // static sequential ids
-pub use {pin::*, res::*, seq::*};
+pub use {marker::*, pin::*, res::*, seq::*};
 
 pub(crate) mod all {
     #[doc(inline)]
     #[allow(unused_imports)]
-    pub use super::{pin::*, res::*, seq::*};
+    pub use super::{marker::*, pin::*, res::*, seq::*};
 }